# be/app/uber/src/grpc-gw/match.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "match.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "MatchService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/estimation": {
      "post": {
        "operationId": "MatchService_estimate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/matchEstimateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/matchEstimateRequest"
            }
          }
        ],
        "tags": [
          "MatchService"
        ]
      }
    },
    "/api/v1/matches": {
      "get": {
        "operationId": "MatchService_ListMatches",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/matchListMatchesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "parent",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageToken",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "MatchService"
        ]
      },
      "post": {
        "operationId": "MatchService_match",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/matchMatch"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/matchMatchRequest"
            }
          }
        ],
        "tags": [
          "MatchService"
        ]
      },
      "patch": {
        "operationId": "MatchService_updateMatch",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/matchMatch"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/matchUpdateMatchRequest"
            }
          }
        ],
        "tags": [
          "MatchService"
        ]
      }
    }
  },
  "definitions": {
    "locationLocation": {
      "type": "object",
      "properties": {
        "lat": {
          "type": "string"
        },
        "lng": {
          "type": "string"
        }
      }
    },
    "matchEstimateRequest": {
      "type": "object",
      "properties": {
        "parent": {
          "type": "string"
        },
        "startingLocation": {
          "type": "string"
        },
        "destination": {
          "type": "string"
        }
      }
    },
    "matchEstimateResponse": {
      "type": "object",
      "properties": {
        "estimation": {
          "$ref": "#/definitions/matchEstimation"
        }
      }
    },
    "matchEstimation": {
      "type": "object",
      "properties": {
        "estimationId": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "startingLocation": {
          "$ref": "#/definitions/locationLocation"
        },
        "desitnation": {
          "$ref": "#/definitions/locationLocation"
        },
        "createdAt": {
          "type": "string"
        },
        "eta": {
          "type": "string"
        },
        "fare": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "title": "Estimation: storage model"
    },
    "matchListMatchesResponse": {
      "type": "object",
      "properties": {
        "matches": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/matchMatch"
          }
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "pageToken": {
          "type": "string"
        }
      }
    },
    "matchMatch": {
      "type": "object",
      "properties": {
        "matchId": {
          "type": "string",
          "format": "int64"
        },
        "driverId": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "estimationId": {
          "type": "string",
          "format": "int64"
        },
        "status": {
          "type": "string"
        }
      },
      "title": "Match: storage model"
    },
    "matchMatchRequest": {
      "type": "object",
      "properties": {
        "parent": {
          "type": "string"
        },
        "estimationId": {
          "type": "string"
        }
      }
    },
    "matchUpdateMatchRequest": {
      "type": "object",
      "properties": {
        "parent": {
          "type": "integer",
          "format": "int32"
        },
        "match": {
          "$ref": "#/definitions/matchMatch"
        },
        "updateMask": {
          "type": "string",
          "title": "repeated"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}